requirements.txt
# Core API
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sse-starlette==2.2.1
httpx==0.28.1

# Data parsing
openpyxl==3.1.5
pdfminer.six==20231228
beautifulsoup4==4.12.3

# Numeric / retrieval
numpy

# Cache & Database
diskcache==5.6.3
aiosqlite==0.20.0

# Utilities
pydantic-settings==2.7.1

# Optional: better embeddings (requires torch runtime)
# sentence-transformers==3.3.1

# Optional: Chroma vector DB backend
# chromadb==0.5.20

# Optional: legacy .xls parser fallback
# pandas>=2.2
# xlrd>=2.0

# Optional: local LLM backend (requires build tools on Windows)
# pip install llama-cpp-python --prefer-binary
